Added spin function for slot machine game
Added a 3 symbol, 5 line slot machine game's spin function. The machine is able to simulate additional lines by "cycling" through what symbols would appear above, below, and diagonal relative to the center line. Each column of the machine does not have to contain the symbols in the same order or the same symbols, it is customizable from the config file.